+++ services/library/internal/interfaces.go
@@ -6,11 +6,11 @@ import (
 )
 
 type LibraryUsecase interface {
-	GetLibrariesList(page, size int, city string) ([]models2.LibraryPaginationResponse, models2.StatusCode)
-	GetBooksList(page, size int, showAll bool, LibUid uuid.UUID) ([]models2.LibraryBookPaginationResponse, models2.StatusCode)
+	GetLibrariesList(page, size int64, city string) ([]models2.LibraryPaginationResponse, models2.StatusCode)
+	GetBooksList(page, size int64, showAll bool, LibUid uuid.UUID) ([]models2.LibraryBookPaginationResponse, models2.StatusCode)
 }
 
 type LibraryRepo interface {
-	GetLibraries(page, size int, city string) ([]models2.LibraryResponse, int, models2.StatusCode)
-	GetBooks(page, size int, showAll bool, LibUid uuid.UUID) ([]models2.LibraryBookResponse, int, models2.StatusCode)
+	GetLibraries(page, size int64, city string) ([]models2.LibraryResponse, int64, models2.StatusCode)
+	GetBooks(page, size int64, showAll bool, LibUid uuid.UUID) ([]models2.LibraryBookResponse, int64, models2.StatusCode)
 }
