+++ services/models/pagination_response.go
@@ -1,15 +1,15 @@
 package models
 
 type LibraryPaginationResponse struct {
-	Page          int               `json:"page"`
-	PageSize      int               `json:"page_size"`
-	TotalElements int               `json:"total_elements"`
+	Page          int64             `json:"page"`
+	PageSize      int64             `json:"page_size"`
+	TotalElements int64             `json:"total_elements"`
 	Items         []LibraryResponse `json:"items"`
 }
 
 type LibraryBookPaginationResponse struct {
-	Page          int                   `json:"page"`
-	PageSize      int                   `json:"page_size"`
-	TotalElements int                   `json:"total_elements"`
+	Page          int64                 `json:"page"`
+	PageSize      int64                 `json:"page_size"`
+	TotalElements int64                 `json:"total_elements"`
 	Items         []LibraryBookResponse `json:"items"`
 }
